Add files via upload
new, updated and tidy version of the code. the clustering was leading to issues on some machines as each version would not be unique. therefore a manual approach was implemented for simplicity and easier debugging. this worked much better; a functioning model was able to run from this script, which was an improvement on the others from yesterday. the next step is to implement advection and add residual norm testing.